+++ ts/features/payments/onboarding/components/WalletPaymentMethodItemSkeleton.tsx
